+++ app/js/main.js
@@ -1,1 +1,4 @@
 console.log('hello world');
+jQuery(document).ready(function($) {
+	console.log('ready');
+});